requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.5.3
asttokens==3.0.0
attrs==23.2.0
Beaker==1.12.1
beautifulsoup4==4.12.3
BitVector==3.5.0
blivet==3.10.0
blivet-gui==2.6.0
boto3==1.36.22
botocore==1.36.22
Brlapi==0.8.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==41.0.7
cson==0.8
cssselect==1.1.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnf==4.22.0
executing==2.1.0
fastapi==0.115.12
fb-re2==1.0.7
fedora-third-party==0.10
file-magic==0.4.0
fonttools==4.51.0
fros==1.1
h11==0.14.0
httpie==3.2.2
humanize==3.13.1
idna==3.7
inkex==1.4.0
ipykernel==6.29.5
ipython==8.31.0
iso639==0.1.4
jedi==0.19.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
langtable==0.0.68
libcomps==0.1.20
libdnf==0.73.4
louis==3.28.0
lxml==5.1.0
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mercurial==6.7.4
mod-wsgi==5.0.0
multidict==6.0.5
nest-asyncio==1.6.0
nftables==0.1
numpy==1.26.4
olefile==0.47
opencv-python-headless==4.11.0.86
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
parso==0.8.4
Paste==3.7.1
pexpect==4.9.0
pid==2.2.3
pillow==10.3.0
platformdirs==4.3.6
ply==3.11
productmd==1.43
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.11.1
pydantic_core==2.33.0
pyenchant==3.2.2
Pygments==2.17.2
PyGObject==3.48.2
pykickstart==3.52
pyOpenSSL==23.2.0
pyparsing==3.1.2
pyparted==3.13.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.13
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.51
python-pam==2.0.2
pytz==2024.2
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==26.2.0
regex==2024.9.11
requests==2.31.0
requests-file==2.0.0
requests-ftp==0.3.1
requests-toolbelt==1.0.0
rich==13.7.0
rpm==4.19.1.1
s3transfer==0.11.2
scour==0.38.2
selenium==4.19.0
selinux @ file:///builddir/build/BUILD/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.0.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.7.2
soupsieve==2.5
speg==0.3
stack-data==0.6.3
starlette==0.46.1
systemd-python==235
Tempita==0.5.2
tinycss2==1.2.1
tornado==6.4.2
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
wsproto==1.2.0
xkbregistry==0.3
